Undid some changes to local copy of oauth2app which forced changes to apps/client/views. Inserted the redirect_uri back. Added instructions to not show sql commands.

+++ myclient/apps/client/views.py
@@ -1,7 +1,9 @@
 #-*- coding: utf-8 -*-
-
+"""
+Interact with the client. 
+"""
 from django.shortcuts import render_to_response
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponse 
 from django.views.decorators.csrf import csrf_exempt
 from django.conf import settings 
 from django.template import RequestContext
@@ -14,10 +16,20 @@ import simplejson as json
 import requests
 import traceback 
 from urllib import urlencode 
+import logging 
 
-"""
-Interact with the client. 
-"""
+#=================================================================
+# Prevent sql commands from being printed out for this section...
+from django.conf import settings
+from django.db.backends import BaseDatabaseWrapper
+from django.db.backends.util import CursorWrapper
+
+if settings.DEBUG:
+    BaseDatabaseWrapper.make_debug_cursor = lambda self, cursor: CursorWrapper(cursor, self)
+#==================================================================
+
+log = logging.getLogger("myclient.apps.client.views") 
+
 @login_required 
 def client(request, client_id):
     """
@@ -27,7 +39,7 @@ def client(request, client_id):
     # XXX The client should eventually be decoupled from the user
     # account information.
 
-    print "In apps.client.views.client - entered" 
+    log.debug("In apps.client.views.client - entered")
     client = Client.objects.get(key=client_id)
     template = {
         "client":client,
@@ -54,19 +66,29 @@ def request(request, req_id):
 
     # XXX Rationalize the URL. Right now it is client/request/aaa
     # it could easily be request/aaa or callback/aaa 
-    print "In apps.client.views.request - entered" 
+    log.debug("In apps.client.views.request - entered")
+    if request.META.has_key('HTTP_REFERRER'):
+        log.debug("Referrer = %s " % (request.META['HTTP_REFERRER']))
 
-    # The oauthRequest object is the state object for the request. It
-    # keeps track of what the user asked and is updated when new
-    # information is received from the server. Lookup that object
-    # first.
-    req = OauthRequest.objects.get(key=req_id)
-    print "Found request = ", vars(req)
     try: 
+
+        # The oauthRequest object is the state object for the request. It
+        # keeps track of what the user asked and is updated when new
+        # information is received from the server. Lookup that object
+        # first.
+        req = OauthRequest.objects.get(key=req_id)
+        log.debug("Found request = %s " % vars(req))
         
         # Find the code 
         code_key = request.GET.get('code')
-        print "Received code = ", code_key
+        log.debug("Received code = %s " % code_key)
+        
+        # => Check where the code has been used. Code object has been 
+        # deleted. So it cant be used. This is a hack to make things
+        # work with oauth2app 
+        if ((req.code == code_key) and (req.completed != 0)): 
+            log.error("Duplicate callback. Should not come here") 
+            raise Exception("Code already used") 
         
         #=> Check if the code exists.
         try: 
@@ -93,14 +115,14 @@ def request(request, req_id):
         code.scope.add(*req.scope.all())
         code.save() 
 
-        print "saved code = ", vars(code) 
-        print "code scope = ", code.scope 
+        log.debug("saved code = %s " % vars(code))
+        log.debug("code scope = %s " % code.scope)
         # XXX This is not being used. May be it can be removed. 
         req.completed = TimestampGenerator()() 
         req.code = code_key # we dont store code object because it
                             # will be deleted
         req.save()
-        print "saved request = ", vars(req) 
+        log.debug("saved request = %s " % vars(req))
 
     except:
         traceback.print_exc() 
@@ -129,10 +151,10 @@ def get_auth(auth_type='basic'):
                         
     client_key = settings.RESOURCE_CLIENT_KEY
     client_secret = settings.RESOURCE_CLIENT_SECRET
-    print "Client_key = ", client_key
-    print "Client secret=", client_secret 
+    log.debug("Client_key = %s " % client_key)
+    log.debug("Client secret= %s" % client_secret)
     basic_auth = "Basic %s" % b64encode(client_key + ":" + client_secret)
-    print "computed authorization = ", basic_auth
+    log.debug("computed authorization = %s " % basic_auth)
     return basic_auth 
 
 @login_required 
@@ -142,7 +164,7 @@ def request_token(request):
     Handle the request from the user to obtain an access token for a
     given code.
     """ 
-    print "Trying to obtain a token" 
+    log.debug("Trying to obtain a token")
     if request.method != "GET": 
         return HttpResponseRedirect("/")
     
@@ -156,81 +178,94 @@ def request_token(request):
         code = Code.objects.get(key=params['code'])
         req = OauthRequest.objects.get(code=params['code'])
         client = req.client 
+
         #client = Client.objects.all()[:1][0] 
-        print "Client object = ", vars(client) 
-        print "Code object = ", vars(code) 
+        log.debug("Client object = %s " % vars(client))
+        log.debug("Code object = %s " % vars(code))
         
         # set the standard parameters 
         params['client_id'] = settings.RESOURCE_CLIENT_KEY         
-        params['grant_type'] = 'authorization_code'         
-        
-        # Ignore the redirect_uri because we dont need a call back. The 
-        # json response is enough. 
-        #params['redirect_uri'] = settings.CLIENT_SITE + code.redirect_uri
-        
+        params['grant_type'] = 'authorization_code'                 
+        params['redirect_uri'] = \
+            "http://%s/client/request/%s" % (request.get_host(), req.key)
+
         # XXX set the scope. Not sure if this the best way. 
-        print "type req.scope.all() = ", type(req.scope.all())
+        log.debug("type req.scope.all() = %s " % type(req.scope.all()))
         all_keys = [s.key for s in req.scope.all()]
-        print "all_scopes = ", all_keys
+        log.debug("all_scopes = %s " % all_keys)
         if len(all_keys) > 0: 
             params['scope'] = all_keys[0] # XXX should be a join(",") instead?
         else:
             params['scope'] = ""
 
-        print "Sending data = ", params         
+        log.debug("Sending data = %s " % params)
         
         # Obtain the authentication 
         basic_auth = get_auth() 
         headers = { 'Authorization': basic_auth } 
-        print "headers = ", headers         
+        log.debug("headers = %s " % headers)
 
         # Constructing the call 
         url = settings.ACCESS_TOKEN_URL + "/?" + urlencode(params)
-        print "url = ", url 
+        log.debug("url = %s " % url)
 
         # There is nothing in the body. There is only a post - which
         # for some reason seems to turn into a GET at the other end. 
         r = requests.post(url, data="", headers=headers)
-        print "received headers = ", r.headers 
+        log.debug("received headers = %s " % r.headers)
         if r.headers['content-type'] != 'application/json':
+            log.error("Received error from server %s" % r.content)
             raise Exception("Possible error in request to the server")
         
         #=> Now store the token for future access purposes 
         grant_data = json.loads(r.content)
-        print "grant data = ", grant_data
+        log.debug("grant data = %s " % grant_data)
         
-        # Create an access token.  There is no place, it seems for a
-        # token_type (bearer etc.)
-        access_token = \
-            AccessToken.objects.create(user=request.user, 
+        content = grant_data 
+        if not grant_data.has_key('error'): 
+            # Create an access token.  There is no place, it seems for a
+            # token_type (bearer etc.)
+            access_token = \
+                AccessToken.objects.create(user=request.user, 
                                        client=client, 
                                        token=grant_data['access_token'],
                                        refresh_token=grant_data['refresh_token'],
                                        # token_type=gr['token_type'],
                                        # scope=grant_data['scope']
                                        )
-        # Should be this [grant_data['scope']]? 
-        access_ranges = list(AccessRange.objects.filter(key__in=[grant_data['scope']]))
-        access_token.scope = access_ranges
-        # alternative to the above
-        # access_token.scope.add(*req.scope.all())
-        access_token.save() 
+            # Should be this [grant_data['scope']]? 
+            access_ranges = list(AccessRange.objects.filter(key__in=[grant_data['scope']]))
+            access_token.scope = access_ranges
+            # alternative to the above
+            # access_token.scope.add(*req.scope.all())
+            access_token.save() 
         
-        # Update the state 
-        req.refresh_token = grant_data['refresh_token'] 
-        req.token = grant_data['access_token'] 
-        req.save() 
-        print "access token = ", vars(access_token) 
+            # Update the state 
+            req.refresh_token = grant_data['refresh_token'] 
+            req.token = grant_data['access_token'] 
+            req.save() 
+            log.debug("access token = %s " % vars(access_token))
         
-        # Clean up 
-        code.delete() 
-        
+            # Clean up 
+            code.delete() 
+            log.debug("Deleted code") 
+
+        # wrap if necessary...
+        try: 
+            callback=request.GET.get('callback')
+            json_data = "%s(%s);" % (callback, json.dumps(content))
+        except: 
+            json_data = json.dumps(content)
+
+        # => Send the response back 
+        log.debug("Response = %s " % json_data) 
+        response = HttpResponse(json_data, 
+                                content_type="application/json")
+        return response         
     except:
         traceback.print_exc() 
         pass
     
-    # redirect 
-    HttpResponseRedirect(settings.CLIENT_SITE + "/client/%s" % client.key)
 
 
 @login_required 
@@ -247,6 +282,8 @@ def refresh_token(request):
     # <input type="hidden" name="scope" value="" />
     # <input type="submit" value="38d6122e30"/>
     # </form>
+    
+    log.debug("in refresh_token")
 
     if request.method != "GET": 
         return HttpResponseRedirect("/")
@@ -257,69 +294,94 @@ def refresh_token(request):
         token = AccessToken.objects.get(refresh_token=refresh_token_key) 
         req = OauthRequest.objects.get(refresh_token=refresh_token_key) 
         client = req.client
+    
+        # Start constructing the request that must be sent to the resource
+        # server.
+        params = {} 
+        params['client_id'] = settings.RESOURCE_CLIENT_KEY         
+        params['grant_type'] = 'refresh_token'
+        params['refresh_token'] = token.refresh_token 
+        
+        # Dont need to specify the redirect_uri as we dont need a call
+        # back from the server. Just the json response is enough.
+        params['redirect_uri'] = \
+            "http://%s/client/request/%s" % (request.get_host(), req.key)
+        
+        # => set the scope of the refresh. Not sure why scope is required
+        # again because it has been specific while obtaining the
+        # authorization. The scope cant be any different. 
+        log.debug("type req.scope.all() = %s " % type(req.scope.all()))
+        all_keys = [s.key for s in req.scope.all()]
+        log.debug("all_scopes = %s " % all_keys)
+        if len(all_keys) > 0: 
+            params['scope'] = all_keys[0]
+        else:
+            params['scope'] = ""
+        
+        # => params to ready 
+        log.debug("params = %s " % params)
+    
+        # Obtain the authentication 
+        basic_auth = get_auth() 
+        headers = { 'Authorization': basic_auth } 
+        log.debug("headers = %s " %  headers)
+        
+        # Constructing the call 
+        url = settings.ACCESS_TOKEN_URL + "/?" + urlencode(params)
+        log.debug("url = %s " % url)
+        
+        # Call the server 
+        r = requests.post(url, data="", headers=headers)
+        log.debug("received headers = %s " % r.headers)
+        if r.headers['content-type'] != 'application/json':
+            # Should probably delete the token 
+            log.error("Received error from server %s" % r.content)
+            raise Exception("Possible error in request to the server")
+        
+        # => Now store the token for future access purposes 
+        grant_data = json.loads(r.content)
+        log.debug("grant data = %s " % grant_data)
+    
+        if grant_data.has_key('error'):
+            # Dont change anything
+            content = grant_data 
+        else: 
+            # => Update the token state 
+            token.token = grant_data['access_token']
+            token.refresh_token = grant_data['refresh_token']
+            now =  TimestampGenerator()()
+            token.expire =now + int(grant_data['expire_in'])
+            token.save() 
+            
+            # Update the request state 
+            req.token = grant_data['access_token']
+            req.refresh_token = grant_data['refresh_token']
+            req.save() 
+            
+            # response 
+            content = {'refresh_token': req.refresh_token, 'token': req.token}
+            
+        # wrap if necessary...
+        try: 
+            callback=request.GET.get('callback')
+            json_data = "%s(%s);" % (callback, json.dumps(content))
+        except: 
+            json_data = json.dumps(content)
+
+        # XXX May be this should be the response instead
+        #next = "/client/%s" % client.key
+        #log.debug("Redirecting to %s " % next)
+        #return HttpResponseRedirect(next)
+            
+        # => Send the response back 
+        log.debug("Response = %s " % json_data) 
+        response = HttpResponse(json_data, 
+                                content_type="application/json")
+        return response 
+    
     except: 
         traceback.print_exc() 
         raise Exception("Invalid refresh token") 
-    
-    # Start constructing the request that must be sent to the resource
-    # server.
-    params = {} 
-    params['client_id'] = settings.RESOURCE_CLIENT_KEY         
-    params['grant_type'] = 'refresh_token'
-    params['refresh_token'] = token.refresh_token 
-    
-    # Dont need to specify the redirect_uri as we dont need a call
-    # back from the server. Just the json response is enough.
-    #params['redirect_uri'] = \
-    #    "http://%s/client/request/%s" % (request.get_host(), req.key)
-
-    print "type req.scope.all() = ", type(req.scope.all())
-    #=> set the scope of the refresh. Not sure why scope is required
-    # again because it has been specific while obtaining the
-    # authorization. The scope cant be any different. 
-    all_keys = [s.key for s in req.scope.all()]
-    print "all_scopes = ", all_keys
-    if len(all_keys) > 0: 
-        params['scope'] = all_keys[0]
-    else:
-        params['scope'] = ""
-        
-    print "params = ", params 
-    
-    # Obtain the authentication 
-    basic_auth = get_auth() 
-    headers = { 'Authorization': basic_auth } 
-    print "headers = ", headers 
-    
-    # Constructing the call 
-    url = settings.ACCESS_TOKEN_URL + "/?" + urlencode(params)
-    print "url = ", url 
-
-    # Call the server 
-    r = requests.post(url, data="", headers=headers)
-    print "received headers = ", r.headers 
-    if r.headers['content-type'] != 'application/json':
-        raise Exception("Possible error in request to the server")
-        
-    # => Now store the token for future access purposes 
-    grant_data = json.loads(r.content)
-    print "grant data = ", grant_data
-    
-    # => Update the token state 
-    token.token = grant_data['access_token']
-    token.refresh_token = grant_data['refresh_token']
-    token.expire = token.expire + int(grant_data['expire_in'])
-    token.save() 
-    
-    # Update the request state 
-    req.token = grant_data['access_token']
-    req.refresh_token = grant_data['refresh_token']
-    req.save() 
-
-    next = "/client/%s" % client.key
-    print "Redirecting to", next 
-    return HttpResponseRedirect(next)
-    
 
 @login_required 
 def forward(request): 
@@ -335,8 +397,8 @@ def forward(request):
     # XXX change /client/forward to something else. may be request_code 
     if request.method != "GET": 
         return HttpResponseRedirect(nexturl)
-
-    print "request parameters = ", request.GET 
+    
+    log.debug("request parameters = %s " % request.GET)
     
     # Figure out if this request is aadhaar 
     try: 
@@ -347,7 +409,7 @@ def forward(request):
     except:
         pass
     
-    print "aadhaar = ", aadhaar, type(aadhaar) 
+    log.debug("aadhaar = %s %s " % (aadhaar, type(aadhaar)))
 
     # Some dummy client. It does not matter. 
     # XXX change this if you need to support multiple clients. 
@@ -357,14 +419,14 @@ def forward(request):
 
     user = request.user 
     scope_key = request.GET.get('scope')
-    print "scope key = ", scope_key, type(scope_key) 
+    log.debug("scope key = %s %s " % (scope_key, type(scope_key)))
     if ((scope_key is not None) and (scope_key != "None")):
         access_ranges = list(AccessRange.objects.filter(key__in=[scope_key]))
     else:
         access_ranges = []
 
-    print "access_ranges = ", access_ranges 
-    print [r.key for r in access_ranges]
+    log.debug("access_ranges = %s " % access_ranges)
+    log.debug([r.key for r in access_ranges])
     response_type = 'code'
     
     # Now create the request. 
@@ -372,10 +434,10 @@ def forward(request):
                                       user=user, 
                                       #aadhaar=aadhaar,
                                       response_type=response_type)
-
     req.scope = access_ranges
     req.save() 
-    print "created request object ", req, "with key ", req.key 
+    
+    log.debug("created request object %s with key %s " % (req, req.key))
 
     # Construct the request 
     #params = request.GET.copy()
@@ -392,8 +454,6 @@ def forward(request):
         next = "%s?%s" % (settings.AADHAAR_AUTHORIZE_URL, url)
     else:
         next = "%s?%s" % (settings.AUTHORIZE_URL, url)
-    print "Updated parameters = ", params 
-    print "Redirecting to ", next 
-
-    print "redirecting the request to ", next 
+    log.debug("Updated parameters = %s " % params)
+    log.debug("Redirecting to %s " % next)
     return HttpResponseRedirect(next)
